Revert "Complete API features and add simple frontend"
