+++ .gitignore
@@ -1,5 +1,4 @@
 composer.phar
-/API/vendor/
 /vendor/
 
 # Commit your application's lock file https://getcomposer.org/doc/01-basic-usage.md#commit-your-composer-lock-file-to-version-control
