+++ README.md
@@ -3,10 +3,10 @@ Authentification centralisée
 
 # Objectif
 Ci dessous se trouve la liste des objectifs de cet API, avec leur avancement respectifs
- - [ ] Inscription
+- [ ] Inscription
     - [X] Vérification de l'Email
-    - [X] Renvoyer l'Email de vérification en cas de tentative de login
-    - [X] Annuler l'inscription
+    - [ ] Renvoyer l'Email de vérification en cas de tentative de login
+    - [ ] Annuler l'inscription
 - [X] Connexion
     - [X] Via Email
     - [X] Via Username
@@ -18,14 +18,14 @@ Ci dessous se trouve la liste des objectifs de cet API, avec leur avancement res
     - [ ] Page de modification du mot de passe
     - [ ] Page de modification de l'adresse Email
 - [X] Suppression du compte
- - [ ] Modification du profile
-    - [X] Modification du pseudo
-    - [X] Modification de l'image de profile
- - [ ] Récupération des informations de l'utilisateurs connecté (inclu la vérification d'une connexion active)
+- [ ] Modification du profile
+    - [ ] Modification du pseudo
+    - [ ] Modification de l'image de profile
+- [ ] Récupération des informations de l'utilisateurs connecté (inclu la vérification d'une connexion active)
     - [X] Informations du compte
-    - [X] Information du profile
- - [ ] Interface Utilisateur
-    - [X] Page de connexion
-    - [X] Page d'inscription
-    - [X] Page de modification du profile
-    - [X] Page de modification des permissions (Role)
+    - [ ] Information du profile
+- [ ] Interface Utilisateur
+    - [ ] Page de connexion
+    - [ ] Page d'inscription
+    - [ ] Page de modification du profile
+    - [ ] Page de modification des permissions (Role)
