--- role.html
@@ -1,33 +1,0 @@
-<!DOCTYPE html>
-<html lang="fr">
-<head>
-<meta charset="UTF-8">
-<title>Permissions</title>
-<style>
-body{font-family:Arial;margin:2em;background:#f7f7f7}
-form{background:#fff;padding:1em;border-radius:5px;max-width:300px;margin:auto}
-select,input{display:block;margin-bottom:1em;width:100%;padding:.5em}
-button{padding:.5em 1em}
-</style>
-</head>
-<body>
-<h2>Changement de rôle</h2>
-<form id="roleForm">
-<select name="role">
-<option value="user">Utilisateur</option>
-<option value="admin">Admin</option>
-</select>
-<button type="submit">Mettre à jour</button>
-</form>
-<script>
-document.getElementById('roleForm').addEventListener('submit', async e=>{
- e.preventDefault();
- const data=Object.fromEntries(new FormData(e.target).entries());
- data.action='updateProfile';
- const r=await fetch('API/index.php?action=updateProfile', {method:'PUT',headers:{'Content-Type':'application/json','SESSION_ID':localStorage.getItem('sid')||''},body:JSON.stringify(data)});
- const res=await r.json();
- alert(res.message || (res.success ? 'Rôle mis à jour' : res.error));
-});
-</script>
-</body>
-</html>
